requirements.txt
streamlit
pandas
mysql-connector-python
plotly
python-dotenv

